Fix all solutions to work with new primality database
Replaces text file with array for faster access in isPrime
Allows use of list if needed through putPrimesToList

+++ Euler/src/com/github/vsams14/euler/solver27.java
@@ -41,7 +41,7 @@ public class solver27 {
 		}
 	}
 	
-	private long formula(int a, int b, int n){
+	private int formula(int a, int b, int n){
 		return n*n+a*n+b;
 	}
 }
